requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.7.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
Django==5.1.3
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-private-storage==3.1.1
django-recaptcha==4.1.0
django-two-factor-auth==1.17.0
fontawesomefree==6.6.0
freeze-requirements==0.5.3
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
packaging==24.2
phonenumbers==8.13.53
pillow==11.0.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
requests==2.32.3
sh==2.2.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
